removes abstraction and splits commands into add and edit where edit command holds most of the optional information for product
